refactor(#3037): relocate review/dispatch loopback DTOs to services (backflow 21→15)
Move three serde request-body DTOs out of `crate::server::routes::*` into the
services layer so the `completion_guard` hot-file (and `recovery_engine`) stop
reaching back into `crate::server` over the internal-HTTP loopback. axum
`Json<T>` extraction is location-independent, so the route handlers now reference
the services paths directly — the dependency direction flips to server → services.

Moved DTOs (serde attrs/fields/derives byte-identical):
- `ReviewDecisionBody` (decision_route.rs) → `services::review_decision`
- `SubmitVerdictBody` + `VerdictItem` (verdict_route.rs) → `services::review_decision`
- `UpdateDispatchBody` (dispatches/crud.rs) → `services::dispatches` (new `dtos.rs`)

Consumers updated:
- server handlers: decision_route.rs / verdict_route.rs / crud.rs reference the
  services-rooted types (crud via the `routes::dispatches` re-export `super::`
  to keep the SQL/json route module off the route-SRP gate).
- completion_guard.rs (hot-file): pure import-path re-point only — no logic /
  control-flow / field / value changes (5 `crate::server` refs → 0).
- recovery_engine.rs: path re-point (1 → 0). internal_api.rs / cli::direct: paths.
- review_verdict/mod.rs: dropped now-unused DTO facade re-exports.

Behavior unchanged: identical JSON shape, validation, and handler dispatch.
#3038 decision_route god-function NOT decomposed — DTO move only.

Gates: cargo fmt --check clean; cargo check --lib / --lib --tests clean
(no new warnings); cargo test --lib turn_bridge (139) + dispatches (45) pass;
service_server_backflow baseline 21 → 15 (completion_guard 5→0, recovery_engine
1→0); generate_inventory_docs + check_agent_maintenance_docs freshness passed;
bash scripts/ci-script-checks.sh exit 0 (route-SRP / giant ratchet within
baseline — crud.rs not newly flagged).

Refs #3037
