+++ ARCHITECTURE.md
@@ -551,6 +551,7 @@ src/
 │   │   │   ├── orchestration.rs
 │   │   │   ├── thread_reuse.rs
 │   │   │   └── transport.rs
+│   │   ├── dtos.rs
 │   │   ├── mod.rs
 │   │   ├── outbox_claiming.rs
 │   │   ├── outbox_queue.rs
