+++ scripts/audit_maintainability/baselines/service_server_backflow.json
@@ -3,7 +3,7 @@
   "rule": "service_server_backflow",
   "description": "No-regression baseline for src/services references to crate::server or super::server server-layer modules.",
   "captured_at": "2026-06-08",
-  "total_count": 21,
+  "total_count": 15,
   "files": {
     "src/services/auto_queue/phase_gate_violations.rs": {
       "count": 1
@@ -13,12 +13,6 @@
     },
     "src/services/discord/internal_api.rs": {
       "count": 1
-    },
-    "src/services/discord/recovery_engine.rs": {
-      "count": 1
-    },
-    "src/services/discord/turn_bridge/completion_guard.rs": {
-      "count": 5
     },
     "src/services/dispatched_sessions.rs": {
       "count": 8
